refact, test: added add_role test, moved parts of roles api to a utils module
