+++ taf/api/roles.py
@@ -1,21 +1,21 @@
-from logging import DEBUG, ERROR, INFO
+from logging import DEBUG, ERROR
 import os
-from typing import Dict, List, Optional, Tuple, Union
+from typing import Dict, List, Optional, Tuple
 import click
 from collections import defaultdict
-from functools import partial
 import json
 from pathlib import Path
 from logdecorator import log_on_end, log_on_error, log_on_start
-from tuf.repository_tool import Repository as TUFRepository, Targets
+from taf.api.utils.roles_utils import _role_obj, create_delegations
+from taf.messages import git_commit_message
+from tuf.repository_tool import Targets
 from taf.api.utils.git_utils import check_if_clean, commit_and_push
 from taf.exceptions import TAFError
 from taf.models.converter import from_dict
 from taf.models.types import RolesIterator
-from taf.models.types import Role, TargetsRole
+from taf.models.types import TargetsRole
 from taf.repositoriesdb import REPOSITORIES_JSON_PATH
-from taf.yubikey import get_key_serial_by_id
-from tuf.repository_tool import TARGETS_DIRECTORY_NAME, Metadata
+from tuf.repository_tool import TARGETS_DIRECTORY_NAME
 import tuf.roledb
 import taf.repositoriesdb as repositoriesdb
 from taf.keys import (
@@ -31,14 +31,11 @@ from taf.api.utils.metadata_utils import (
 from taf.auth_repo import AuthenticationRepository
 from taf.constants import (
     DEFAULT_ROLE_SETUP_PARAMS,
-    YUBIKEY_EXPIRATION_DATE,
     DEFAULT_RSA_SIGNATURE_SCHEME,
 )
 from taf.keystore import default_keystore_path, new_public_key_cmd_prompt
 from taf.repository_tool import (
-    Repository,
     is_delegated_role,
-    yubikey_signature_provider,
 )
 from taf.utils import get_key_size, read_input_dict
 from taf.log import taf_logger
@@ -71,6 +68,8 @@ def add_role(
     auth_repo: Optional[AuthenticationRepository] = None,
     commit: Optional[bool] = True,
     prompt_for_keys: Optional[bool] = False,
+    push: Optional[bool] = True,
+    skip_prompt: Optional[bool] = False,
 ) -> None:
     """
     Add a new delegated target role and update and sign metadata files.
@@ -89,6 +88,9 @@ def add_role(
         auth_repo (optional): Instance of the authentication repository. Will be created if not passed into the function.
         commit (optional): Indicates if the changes should be committed and pushed automatically.
         prompt_for_keys (optional): Whether to ask the user to enter their key if it is not located inside the keystore directory.
+        push (optional): Flag specifying whether to push to remote.
+        skip_prompt (optional): A flag defining if the user will be asked if they want to generate new keys or reuse existing
+            ones in case keystore files should be used. New keys will be generated by default.
 
     Side Effects:
         Initializes a new delegated targets role, signs metadata files, write changes to the disk and optionally commits.
@@ -122,6 +124,7 @@ def add_role(
         roles=new_role,
         yubikeys_data=None,
         keystore=keystore,
+        skip_prompt=skip_prompt,
     )
     create_delegations(
         new_role, auth_repo, verification_keys, signing_keys, existing_roles
@@ -137,7 +140,10 @@ def add_role(
         update_snapshot_and_timestamp(
             auth_repo, keystore, scheme=scheme, prompt_for_keys=prompt_for_keys
         )
-        commit_and_push(auth_repo)
+        commit_msg = git_commit_message(
+            "add-role", role=role
+        )
+        commit_and_push(auth_repo, commit_msg=commit_msg, push=push)
     else:
         taf_logger.info("\nPlease commit manually\n")
 
@@ -159,6 +165,7 @@ def add_role_paths(
     auth_repo: Optional[AuthenticationRepository] = None,
     auth_path: Optional[str] = None,
     prompt_for_keys: Optional[bool] = False,
+    push: Optional[bool] = True,
 ) -> None:
     """
     Adds additional delegated target paths to the specified role. That means that
@@ -172,6 +179,7 @@ def add_role_paths(
         auth_repo (optional): Instance of the authentication repository. Will be created if not passed into the function.
         prompt_for_keys (optional): Whether to ask the user to enter their key if it is not located inside the keystore directory.
         commit (optional): Indicates if the changes should be committed and pushed automatically.
+        push (optional): Flag specifying whether to push to remote.
     Side Effects:
         Updates the specified target role's parent and other metadata files (snapshot and timestamp),
         signs them, writes changes to disk and optionally commits everything.
@@ -190,7 +198,10 @@ def add_role_paths(
             update_snapshot_and_timestamp(
                 auth_repo, keystore, prompt_for_keys=prompt_for_keys
             )
-            commit_and_push(auth_repo)
+            commit_msg = git_commit_message(
+                "add-role-paths", paths=", ".join(paths), role=delegated_role
+            )
+            commit_and_push(auth_repo, commit_msg=commit_msg, push=push)
         else:
             taf_logger.info("\nPlease commit manually\n")
     else:
@@ -216,6 +227,7 @@ def add_roles(
     scheme: Optional[str] = DEFAULT_RSA_SIGNATURE_SCHEME,
     prompt_for_keys: Optional[bool] = False,
     commit: Optional[bool] = True,
+    push: Optional[bool] = True,
 ) -> None:
     """
     Add new target roles and sign all metadata files given information stored in roles_key_infos
@@ -229,7 +241,7 @@ def add_roles(
         scheme (optional): Signing scheme. Set to rsa-pkcs1v15-sha256 by default.
         prompt_for_keys (optional): Whether to ask the user to enter their key if it is not located inside the keystore directory.
         commit (optional): Indicates if the changes should be committed and pushed automatically.
-
+        push (optional): Flag specifying whether to push to remote.
     Side Effects:
         Updates metadata files (parent of new roles, snapshot and timestamp) and creates new targets metadata files.
         Writes changes to disk.
@@ -290,7 +302,10 @@ def add_roles(
         auth_repo, keystore, scheme=scheme, prompt_for_keys=prompt_for_keys
     )
     if commit:
-        commit_and_push(auth_repo)
+        commit_msg = git_commit_message(
+            "add-roles", roles=", ".join(new_roles)
+        )
+        commit_and_push(auth_repo, commit_msg=commit_msg, push=push)
     else:
         taf_logger.info("\nPlease commit manually\n")
 
@@ -314,6 +329,8 @@ def add_signing_key(
     scheme: Optional[str] = DEFAULT_RSA_SIGNATURE_SCHEME,
     commit: Optional[bool] = True,
     prompt_for_keys: Optional[bool] = False,
+    push: Optional[bool] = True,
+    commit_msg: Optional[str] = None,
 ) -> None:
     """
     Add a new signing key to the listed roles. Update root metadata if one or more roles is one of the main TUF roles,
@@ -329,7 +346,8 @@ def add_signing_key(
         scheme (optional): Signing scheme. Set to rsa-pkcs1v15-sha256 by default.
         prompt_for_keys (optional): Whether to ask the user to enter their key if it is not located inside the keystore directory.
         commit (optional): Indicates if the changes should be committed and pushed automatically.
-
+        push (optional): Flag specifying whether to push to remote.
+        commit_msg(optional): Commit message. Will be necessary to enter it if not provided.
     Side Effects:
         Updates metadata files (parents of the affected roles, snapshot and timestamp).
         Writes changes to disk.
@@ -384,7 +402,12 @@ def add_signing_key(
     )
 
     if commit:
-        commit_and_push(auth_repo)
+        # TODO after saving custom key ids is implemented, remove customization of the commit message
+        # for now, it might be helpful to be able to specify which key was added
+        # commit_msg = git_commit_message(
+        #     "add-signing-key", role={role}
+        # )
+        commit_and_push(auth_repo, commit_msg=commit_msg, push=push)
     else:
         taf_logger.info("\nPlease commit manually\n")
 
@@ -587,94 +610,10 @@ def _initialize_roles_and_keystore(
     return roles_key_infos_dict, keystore
 
 
-@log_on_start(INFO, "Creating delegations", logger=taf_logger)
-@log_on_end(DEBUG, "Finished creating delegations", logger=taf_logger)
-def create_delegations(
-    role: Role,
-    repository: AuthenticationRepository,
-    verification_keys: Dict,
-    signing_keys: Dict,
-    existing_roles: Optional[List[str]] = None,
-) -> None:
-    """
-    Initialize new delegated target roles, update authentication repository object
-
-    Arguments:
-        role: Targets main role or a delegated role
-        repository: Authentication repository.
-        verification_keys: A dictionary containing mappings of role names to their verification (public) keys.
-        signing_keys: A dictionary containing mappings of role names to their signing (private) keys.
-        existing_roles: A list of already initialized roles.
-
-    Side Effects:
-        Updates authentication repository object
-
-    Returns:
-        None
-    """
-    if existing_roles is None:
-        existing_roles = []
-    skip_top_role = isinstance(role, TargetsRole)
-    for delegated_role in RolesIterator(role, skip_top_role=skip_top_role):
-        parent_role_obj = _role_obj(delegated_role.parent.name, repository)
-        if not isinstance(parent_role_obj, Targets):
-            raise TAFError(
-                f"Could not find parent targets role of role {delegated_role}"
-            )
-        if delegated_role.name in existing_roles:
-            taf_logger.info(f"Role {delegated_role.name} already set up.")
-            continue
-        paths = delegated_role.paths
-        roles_verification_keys = verification_keys[delegated_role.name]
-        # if yubikeys are used for signing, signing keys are not loaded
-        roles_signing_keys = signing_keys.get(delegated_role.name)
-        parent_role_obj.delegate(
-            delegated_role.name,
-            roles_verification_keys,
-            paths,
-            threshold=delegated_role.threshold,
-            terminating=delegated_role.terminating,
-        )
-        setup_role(
-            delegated_role,
-            repository,
-            roles_verification_keys,
-            roles_signing_keys,
-            parent=parent_role_obj,
-        )
-
-
 def _get_roles_key_size(role: str, keystore: str, keys_num: int) -> int:
     pub_key_name = f"{get_key_name(role, 1, keys_num)}.pub"
     key_path = str(Path(keystore, pub_key_name))
     return get_key_size(key_path)
-
-
-def _role_obj(
-    role: str,
-    repository: Union[Repository, TUFRepository],
-    parent: Optional[Targets] = None,
-) -> Metadata:
-    """
-    Return role TUF object based on its name
-    """
-    if isinstance(repository, Repository):
-        tuf_repository = repository._repository
-    else:
-        tuf_repository = repository
-    if role == "targets":
-        return tuf_repository.targets
-    elif role == "snapshot":
-        return tuf_repository.snapshot
-    elif role == "timestamp":
-        return tuf_repository.timestamp
-    elif role == "root":
-        return tuf_repository.root
-    else:
-        # return delegated role
-        if parent is None:
-            return tuf_repository.targets(role)
-        return parent(role)
 
 
 @log_on_error(
@@ -727,6 +666,7 @@ def remove_role(
     remove_targets: Optional[bool] = False,
     auth_repo: Optional[AuthenticationRepository] = None,
     prompt_for_keys: Optional[bool] = False,
+    push: Optional[bool] = True,
 ) -> None:
     """
     Remove a delegated target role and update and sign metadata files.
@@ -743,7 +683,7 @@ def remove_role(
             Set to False by default.
         auth_repo (optional): Instance of the authentication repository. Will be created if not passed into the function.
         prompt_for_keys (optional): Whether to ask the user to enter their key if it is not located inside the keystore directory.
-
+        push (optional): Flag specifying whether to push to remote.
     Side Effects:
         Updates metadata files, optionally deletes target files, writes changes to disk and optionally commits.
 
@@ -820,7 +760,10 @@ def remove_role(
         auth_repo, keystore, scheme=scheme, prompt_for_keys=prompt_for_keys
     )
     if commit:
-        commit_and_push(auth_repo)
+        commit_msg = git_commit_message(
+            "remove-role", role=role
+        )
+        commit_and_push(auth_repo, commit_msg=commit_msg, push=push)
     else:
         taf_logger.info("Please commit manually")
 
@@ -840,6 +783,7 @@ def remove_paths(
     keystore: str,
     commit: Optional[bool] = True,
     prompt_for_keys: Optional[bool] = False,
+    push: Optional[bool] = True,
 ) -> bool:
     """
     Remove delegated paths. Update parent roles of the roles associated with the removed paths,
@@ -851,7 +795,7 @@ def remove_paths(
         keystore: Location of the keystore files.
         commit (optional): Indicates if the changes should be committed and pushed automatically.
         prompt_for_keys (optional): Whether to ask the user to enter their key if it is not located inside the keystore directory.
-
+        push (optional): Flag specifying whether to push to remote.
     Side Effects:
         Updates metadata files, writes changes to disk and optionally commits them.
 
@@ -877,7 +821,10 @@ def remove_paths(
         update_snapshot_and_timestamp(
             auth_repo, keystore, prompt_for_keys=prompt_for_keys
         )
-        commit_and_push(auth_repo)
+        commit_msg = git_commit_message(
+            "remove-role-paths", paths=", ".join(paths), role=delegated_role
+        )
+        commit_and_push(auth_repo, commit_msg=commit_msg, push=push)
     elif delegation_existed:
         taf_logger.info("\nPlease commit manually\n")
     return delegation_existed
@@ -928,42 +875,6 @@ def _remove_path_from_role_info(
             parent_role, roleinfo, repository_name=auth_repo.name
         )
     return delegation_exists
-
-
-@log_on_start(INFO, "Setting up role {role.name:s}", logger=taf_logger)
-@log_on_end(DEBUG, "Finished setting up role {role.name:s}", logger=taf_logger)
-def setup_role(
-    role: Role,
-    repository: Repository,
-    verification_keys: Dict,
-    signing_keys: Optional[Dict] = None,
-    parent: Optional[Targets] = None,
-) -> None:
-    """
-    Initialize a new role, add signing and verification keys.
-    """
-    role_obj = _role_obj(role.name, repository, parent)
-    role_obj.threshold = role.threshold
-    if not role.is_yubikey:
-        if verification_keys is None or signing_keys is None:
-            raise TAFError(f"Cannot setup role {role.name}. Keys not specified")
-        for public_key, private_key in zip(verification_keys, signing_keys):
-            role_obj.add_verification_key(public_key)
-            role_obj.load_signing_key(private_key)
-    else:
-        yubikeys = role.yubikeys
-        if yubikeys is None:
-            yubikeys = [
-                get_key_name(role.name, count, role.number)
-                for count in range(role.number)
-            ]
-        for key_name, key in zip(yubikeys, verification_keys):
-            role_obj.add_verification_key(key, expires=YUBIKEY_EXPIRATION_DATE)
-            # check if yubikey loaded
-            if get_key_serial_by_id(key_name):
-                role_obj.add_external_signature_provider(
-                    key, partial(yubikey_signature_provider, key_name, key["keyid"])
-                )
 
 
 def _update_role(
