+++ taf/keys.py
@@ -92,6 +92,7 @@ def load_sorted_keys_of_new_roles(
     keystore: str,
     yubikeys: Optional[Dict[str, Dict]] = None,
     existing_roles: Optional[List[str]] = None,
+    skip_prompt: Optional[bool] = False,
 ):
     """
     Load signing keys of roles - first those stored on YubiKeys to avoid entering pins
@@ -108,7 +109,8 @@ def load_sorted_keys_of_new_roles(
         keystore: keystore path
         yubikeys:(optional): A dictionary containing previously loaded YubiKeys used to save already entered pins
         existing_roles (optional): A list of roles whose keys were already loaded
-
+        skip_prompt (optional): A flag defining if the user will be asked if they want to generate new keys or reuse existing
+            ones in case keystore files should be used. New keys will be generated by default.
     Side Effects:
         Populates loaded YubiKeys database located in `yubikey.py`
 
@@ -139,7 +141,7 @@ def load_sorted_keys_of_new_roles(
         for role in keystore_roles:
             if role.name in existing_roles:
                 continue
-            keystore_keys, _ = setup_roles_keys(role, auth_repo.path, keystore=keystore)
+            keystore_keys, _ = setup_roles_keys(role, auth_repo.path, keystore=keystore, skip_prompt=skip_prompt)
             for public_key, private_key in keystore_keys:
                 signing_keys.setdefault(role.name, []).append(private_key)
                 verification_keys.setdefault(role.name, []).append(public_key)
@@ -152,6 +154,7 @@ def load_sorted_keys_of_new_roles(
                 certs_dir=auth_repo.certs_dir,
                 yubikeys=yubikeys,
                 users_yubikeys_details=yubikeys_data,
+                skip_prompt=skip_prompt,
             )
             verification_keys[role.name] = yubikey_keys
         return signing_keys, verification_keys
@@ -269,6 +272,7 @@ def setup_roles_keys(
     keystore: Optional[str] = None,
     yubikeys: Optional[Dict] = None,
     users_yubikeys_details: Optional[Dict[str, UserKeyData]] = None,
+    skip_prompt: Optional[bool] = False,
 ):
 
     if role.name is None:
@@ -300,6 +304,7 @@ def setup_roles_keys(
                 role.scheme or default_params["scheme"],
                 role.length or default_params["length"],
                 None,
+                skip_prompt=skip_prompt
             )
             keystore_keys.append((public_key, private_key))
     return keystore_keys, yubikey_keys
@@ -361,6 +366,7 @@ def _setup_keystore_key(
     scheme: str,
     length: int,
     password: Optional[str],
+    skip_prompt: Optional[bool],
 ) -> Tuple[Optional[Dict], Optional[Dict]]:
     # if keystore exists, load the keys
     generate_new_keys = keystore is None
@@ -398,7 +404,7 @@ def _setup_keystore_key(
                 _invalid_key_message(key_name, keystore, False)
 
             if public_key is None or private_key is None:
-                generate_new_keys = click.confirm("Generate new keys?")
+                generate_new_keys = skip_prompt is True or click.confirm("Generate new keys?")
                 if not generate_new_keys:
                     if click.confirm("Reuse existing key?"):
                         reused_key_name = input(
@@ -416,8 +422,8 @@ def _setup_keystore_key(
                 else:
                     break
     if generate_new_keys:
-        if keystore is not None and click.confirm("Write keys to keystore files?"):
-            if password is None:
+        if keystore is not None and (skip_prompt or click.confirm("Write keys to keystore files?")):
+            if password is None and not skip_prompt:
                 password = input(
                     "Enter keystore password and press ENTER (can be left empty)"
                 )
